+++ django_comments_tree/abstract.py
@@ -16,25 +16,38 @@ class CommentAbstractModel(models.Model):
     # Who posted this comment? If ``user`` is set then it was an authenticated
     # user; otherwise at least user_name should have been set and the comment
     # was posted by a non-authenticated user.
-    user = models.ForeignKey(settings.AUTH_USER_MODEL, verbose_name=_('user'),
-                             blank=True, null=True, related_name="%(class)s_comments",
+    user = models.ForeignKey(settings.AUTH_USER_MODEL,
+                             verbose_name=_('user'),
+                             blank=True, null=True,
+                             related_name="%(class)s_comments",
                              on_delete=models.SET_NULL)
-    user_name = models.CharField(_("user's name"), max_length=50, blank=True)
-    user_email = models.EmailField(_("user's email address"), blank=True)
+    user_name = models.CharField(_("user's name"),
+                                 max_length=50, blank=True)
+    user_email = models.EmailField(_("user's email address"),
+                                   blank=True)
     user_url = models.URLField(_("user's URL"), blank=True)
 
     comment = MarkupField(_('comment'),
                           default_markup_type='Draft.js')
 
     # Metadata about the comment
-    submit_date = models.DateTimeField(_('date/time submitted'), db_index=True,
+    submit_date = models.DateTimeField(_('date/time submitted'),
+                                       db_index=True,
                                        default=timezone.now)
+
+    updated_on = models.DateTimeField(_('date/time updated'),
+                                      db_index=True,
+                                      default=timezone.now)
+
     ip_address = models.GenericIPAddressField(_('IP address'),
-                                              unpack_ipv4=True, blank=True, null=True)
+                                              unpack_ipv4=True,
+                                              blank=True,
+                                              null=True)
     is_public = models.BooleanField(_('is public'), default=True,
                                     help_text=_('Uncheck this box to make the comment effectively '
                                                 'disappear from the site.'))
-    is_removed = models.BooleanField(_('is removed'), default=False,
+    is_removed = models.BooleanField(_('is removed'),
+                                     default=False,
                                      help_text=_('Check this box if the comment is inappropriate. '
                                                  'A "This comment has been removed" message will '
                                                  'be displayed instead.'))