+++ django_comments_tree/forms.py
@@ -15,6 +15,7 @@ class TreeCommentForm(CommentForm):
                                   widget=forms.HiddenInput())
 
     def __init__(self, *args, **kwargs):
+
         comment = kwargs.pop("comment", None)
         if comment:
             initial = kwargs.pop("initial", {})
@@ -45,6 +46,43 @@ class TreeCommentForm(CommentForm):
         self.fields['comment'].widget.attrs.pop('rows')
         self.fields['followup'].widget.attrs['id'] = (
             'id_followup%s' % followup_suffix)
+        self.fields['followup'].widget.attrs['class'] = "custom-control-input"
+
+    def init_from_data(self, *args, data=None, **kwargs):
+        """ Form used from API calls """
+
+        comment = data.pop("comment", None)
+        if comment:
+            initial = data.pop("initial", {})
+            initial.update({"reply_to": comment.pk})
+            data["initial"] = initial
+            followup_suffix = ('_%d' % comment.pk)
+        else:
+            followup_suffix = ''
+
+        super().__init__(*args, **data)
+
+        self.fields['name'] = forms.CharField(
+            label=_("Name"),
+            widget=forms.TextInput(attrs={'placeholder': _('name'),
+                                          'class': 'form-control'}))
+        self.fields['email'] = forms.EmailField(
+            label=_("Mail"), help_text=_("Required for comment verification"),
+            widget=forms.TextInput(attrs={'placeholder': _('mail address'),
+                                          'class': 'form-control'}))
+        self.fields['url'] = forms.URLField(
+            label=_("Link"), required=False,
+            widget=forms.TextInput(attrs={
+                'placeholder': _('url your name links to (optional)'),
+                'class': 'form-control'}))
+        self.fields['comment'] = forms.CharField(
+            widget=forms.Textarea(attrs={'placeholder': _('Your comment'),
+                                         'class': 'form-control'}),
+            max_length=settings.COMMENT_MAX_LENGTH)
+        self.fields['comment'].widget.attrs.pop('cols')
+        self.fields['comment'].widget.attrs.pop('rows')
+        self.fields['followup'].widget.attrs['id'] = (
+                'id_followup%s' % followup_suffix)
         self.fields['followup'].widget.attrs['class'] = "custom-control-input"
 
     def get_comment_model(self):